test: enforce provider model type ordering
